% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{friendlynumber_default_options}
\alias{friendlynumber_default_options}
\title{Get the default options set by the friendlynumber package}
\usage{
friendlynumber_default_options()
}
\value{
A named list of options.
}
\description{
Returns a list of options provided to \code{options()} when the friendlynumber
package is loaded. Options set prior to loading the friendlynumber package
are not overwritten on load.
}
\examples{
friendlynumber_default_options()

}
